fix(stats): correct import average-time + widen counters to 64-bit (#27)
Three statistics-integrity fixes from the 2026-05-25 code review:

- importStats computed each difficulty's average completion time after it
  had already folded the imported count into games_completed, weighting the
  local term by the post-merge total and dividing by the same inflated
  total. The average drifted on every import and was unrepairable. It is now
  a weighted mean of the two pre-merge datasets.

- Cumulative AggregateStats counters (total_games/completed/moves/hints/
  mistakes, per-difficulty games_played/completed, both win streaks) were
  32-bit int and could wrap at 2^31 for long-lived players. Widened to
  int64_t, with matching StatsDisplay fields and .as<int64_t>() reads. The
  aggregate_stats.yaml schema is unchanged (YAML integers), so existing
  files load unchanged.

- The reported recalculateAggregateStats double-count does not occur: the
  cache is reset before re-folding and the only in-instance trigger is
  private and unused. Reading disk-only (the issue's literal suggestion)
  would instead drop unflushed pending sessions. Added an explanatory
  comment plus a white-box regression test (via a friend test-peer) pinning
  "each disk+pending session counted exactly once".

Full suite green: unit 1053, integration 14, UI 13.

+++ src/core/i_statistics_manager.h
@@ -46,9 +46,9 @@ struct GameStats {
 
 /// Aggregate statistics across multiple games
 struct AggregateStats {
-    // Per difficulty level
-    std::array<int, DIFFICULTY_COUNT> games_played{0, 0, 0, 0, 0};  // Easy, Medium, Hard, Expert, Master
-    std::array<int, DIFFICULTY_COUNT> games_completed{0, 0, 0, 0, 0};
+    // Per difficulty level. 64-bit so lifetime cumulative counts can't wrap (issue #27).
+    std::array<int64_t, DIFFICULTY_COUNT> games_played{0, 0, 0, 0, 0};  // Easy, Medium, Hard, Expert, Master
+    std::array<int64_t, DIFFICULTY_COUNT> games_completed{0, 0, 0, 0, 0};
     std::array<std::chrono::milliseconds, DIFFICULTY_COUNT> best_times{
         std::chrono::milliseconds::max(), std::chrono::milliseconds::max(), std::chrono::milliseconds::max(),
         std::chrono::milliseconds::max(), std::chrono::milliseconds::max()};
@@ -63,17 +63,17 @@ struct AggregateStats {
         std::numeric_limits<double>::max(), std::numeric_limits<double>::max()};
     std::array<double, DIFFICULTY_COUNT> max_ratings{0.0, 0.0, 0.0, 0.0, 0.0};
 
-    // Overall stats
-    int total_games{0};
-    int total_completed{0};
-    int total_moves{0};
-    int total_hints{0};
-    int total_mistakes{0};
+    // Overall stats. 64-bit to avoid 2^31 overflow for long-lived players (issue #27).
+    int64_t total_games{0};
+    int64_t total_completed{0};
+    int64_t total_moves{0};
+    int64_t total_hints{0};
+    int64_t total_mistakes{0};
     std::chrono::milliseconds total_time_played{0};
 
     // Streaks
-    int current_win_streak{0};
-    int best_win_streak{0};
+    int64_t current_win_streak{0};
+    int64_t best_win_streak{0};
 
     // First played
     std::chrono::system_clock::time_point first_game_date;
